Move testAutoStartStop to under FrontEndTest class
