+++ src/test/java/io/antmedia/integration/StreamFetcherV2Test.java
@@ -1,14 +1,12 @@
 package io.antmedia.integration;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 import java.io.File;
 import java.io.IOException;
-import java.time.Duration;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
@@ -24,9 +22,6 @@ import org.junit.rules.TestRule;
 import org.junit.rules.TestWatcher;
 import org.junit.runner.Description;
 import org.mockito.Mockito;
-import org.openqa.selenium.By;
-import org.openqa.selenium.StaleElementReferenceException;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.red5.server.scheduling.QuartzSchedulingService;
 import org.red5.server.scope.WebScope;
 import org.slf4j.Logger;
@@ -44,8 +39,6 @@ import io.antmedia.datastore.db.types.Broadcast;
 import io.antmedia.datastore.db.types.Endpoint;
 import io.antmedia.rest.model.Result;
 import io.antmedia.streamsource.StreamFetcher;
-import io.antmedia.test.StreamFetcherUnitTest;
-import io.github.bonigarcia.wdm.WebDriverManager;
 import io.vertx.core.Vertx;
 
 @ContextConfiguration(locations = { "../test/test.xml" })
@@ -109,7 +102,6 @@ public class StreamFetcherV2Test extends AbstractJUnit4SpringContextTests{
 		//	avformat.av_register_all();
 		avformat.avformat_network_init();
 		avutil.av_log_set_level(avutil.AV_LOG_INFO);
-		WebDriverManager.chromedriver().setup();
 
 	}
 
@@ -161,71 +153,6 @@ public class StreamFetcherV2Test extends AbstractJUnit4SpringContextTests{
 		return appSettings;
 	}
 
-
-	@Test
-	public void testAutoStartStop() {
-		
-		RestServiceV2Test restService = new RestServiceV2Test();
-		StreamFetcherUnitTest.startCameraEmulator();
-
-		Broadcast broadcast = new Broadcast("rtsp_source", null, null, null, "rtsp://127.0.0.1:6554/test.flv",
-				AntMediaApplicationAdapter.STREAM_SOURCE);
-		broadcast.setAutoStartStopEnabled(true);
-		
-		Broadcast streamSource = restService.createBroadcast(broadcast);
-
-		assertNotNull(streamSource);
-		assertEquals(broadcast.getStreamUrl(), streamSource.getStreamUrl());
-		
-		Awaitility.await().pollDelay(10, TimeUnit.SECONDS).atMost(20, TimeUnit.SECONDS).pollInterval(1, TimeUnit.SECONDS).until(() -> {
-			Broadcast localBroadcast = restService.getBroadcast(streamSource.getStreamId());
-			
-			return localBroadcast.getStatus().equals(AntMediaApplicationAdapter.BROADCAST_STATUS_CREATED);
-		});
-		
-		//make a request to play the stream
-		
-		ChromeDriver driver = new ChromeDriver(FrontEndTest.getChromeOptions());
-		driver.manage().timeouts().pageLoadTimeout(Duration.ofSeconds(10));
-		driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
-		
-		String url = "http://localhost:5080/LiveApp/";
-
-		//get with default code & it should fallback to hls and play
-		driver.get(url+"play.html?id="+streamSource.getStreamId() + "&playOrder=hls");
-		
-		
-		//check that it's started to play
-
-		Awaitility.await().atMost(20, TimeUnit.SECONDS).pollInterval(1, TimeUnit.SECONDS).until(() -> {
-
-			try {
-				String readyState = driver.findElement(By.tagName("video")).getDomProperty("readyState");
-				//this.driver.findElement(By.xpath("//*[@id='video-player']")).
-				logger.info("player ready state -> {}", readyState);
-	
-				return readyState != null && readyState.equals("4");
-			}
-			catch (StaleElementReferenceException e) {
-				logger.warn("Stale element exception");
-				return false;
-			}
-		});
-		
-		//stop the player
-		driver.quit();
-		
-		//check that it's stopped because it's hls it takes more time to understand there is no viewer
-		Awaitility.await().atMost(45, TimeUnit.SECONDS).pollInterval(3, TimeUnit.SECONDS).until(() -> {
-			Broadcast localBroadcast = restService.getBroadcast(streamSource.getStreamId());
-			
-			return localBroadcast.getStatus().equals(AntMediaApplicationAdapter.BROADCAST_STATUS_FINISHED);
-		});
-		
-		
-		StreamFetcherUnitTest.stopCameraEmulator();
-		
-	}
 
 	@Test
 	public void testUpdateStreamSource() {
